+++ source/extensions/filters/http/lua/lua_filter.h
@@ -74,6 +74,11 @@ public:
    * accomodate write API e.g. setDynamicMetadata().
    */
   virtual RequestInfo::RequestInfo& requestInfo() PURE;
+
+  /**
+   * @return const const Network::Connection* the current network connection handle.
+   */
+  virtual const Network::Connection* connection() const PURE;
 };
 
 class Filter;
@@ -128,7 +133,8 @@ public:
             {"logDebug", static_luaLogDebug},       {"logInfo", static_luaLogInfo},
             {"logWarn", static_luaLogWarn},         {"logErr", static_luaLogErr},
             {"logCritical", static_luaLogCritical}, {"httpCall", static_luaHttpCall},
-            {"respond", static_luaRespond},         {"requestInfo", static_luaRequestInfo}};
+            {"respond", static_luaRespond},         {"requestInfo", static_luaRequestInfo},
+            {"connection", static_luaConnection}};
   }
 
 private:
@@ -188,6 +194,11 @@ private:
   DECLARE_LUA_FUNCTION(StreamHandleWrapper, luaRequestInfo);
 
   /**
+   * @return a handle to the network connection.
+   */
+  DECLARE_LUA_FUNCTION(StreamHandleWrapper, luaConnection);
+
+  /**
    * Log a message to the Envoy log.
    * @param 1 (string): The log message.
    */
@@ -214,6 +225,7 @@ private:
     trailers_wrapper_.reset();
     metadata_wrapper_.reset();
     request_info_wrapper_.reset();
+    connection_wrapper_.reset();
   }
 
   // Http::AsyncClient::Callbacks
@@ -232,8 +244,9 @@ private:
   Filters::Common::Lua::LuaDeathRef<HeaderMapWrapper> headers_wrapper_;
   Filters::Common::Lua::LuaDeathRef<Filters::Common::Lua::BufferWrapper> body_wrapper_;
   Filters::Common::Lua::LuaDeathRef<HeaderMapWrapper> trailers_wrapper_;
-  Filters::Common::Lua::LuaDeathRef<RequestInfoWrapper> request_info_wrapper_;
   Filters::Common::Lua::LuaDeathRef<Filters::Common::Lua::MetadataMapWrapper> metadata_wrapper_;
+  Filters::Common::Lua::LuaDeathRef<RequestInfoWrapper> request_info_wrapper_;
+  Filters::Common::Lua::LuaDeathRef<Filters::Common::Lua::ConnectionWrapper> connection_wrapper_;
   State state_{State::Running};
   std::function<void()> yield_callback_;
   Http::AsyncClient::Request* http_request_{};
@@ -324,6 +337,7 @@ private:
 
     const ProtobufWkt::Struct& metadata() const override { return getMetadata(callbacks_); }
     RequestInfo::RequestInfo& requestInfo() override { return callbacks_->requestInfo(); }
+    const Network::Connection* connection() const override { return callbacks_->connection(); }
 
     Filter& parent_;
     Http::StreamDecoderFilterCallbacks* callbacks_{};
@@ -343,6 +357,7 @@ private:
 
     const ProtobufWkt::Struct& metadata() const override { return getMetadata(callbacks_); }
     RequestInfo::RequestInfo& requestInfo() override { return callbacks_->requestInfo(); }
+    const Network::Connection* connection() const override { return callbacks_->connection(); }
 
     Filter& parent_;
     Http::StreamEncoderFilterCallbacks* callbacks_{};