+++ test/extensions/filters/common/lua/wrappers_test.cc
@@ -3,6 +3,8 @@
 #include "extensions/filters/common/lua/wrappers.h"
 
 #include "test/extensions/filters/common/lua/lua_wrappers.h"
+#include "test/mocks/network/mocks.h"
+#include "test/mocks/ssl/mocks.h"
 #include "test/test_common/utility.h"
 
 namespace Envoy {
@@ -25,6 +27,42 @@ public:
     MessageUtil::loadFromYaml(yaml_string, metadata);
     return metadata;
   }
+};
+
+class LuaConnectionWrapperTest : public LuaWrappersTestBase<ConnectionWrapper> {
+public:
+  virtual void setup(const std::string& script) {
+    LuaWrappersTestBase<ConnectionWrapper>::setup(script);
+    state_->registerType<SslConnectionWrapper>();
+  }
+
+protected:
+  void expectSecureConnection(const bool secure) {
+    const std::string SCRIPT{R"EOF(
+      function callMe(object)
+        if object:ssl() == nil then
+          testPrint("plain")
+        else
+          testPrint("secure")
+        end
+        testPrint(type(object:ssl()))
+      end
+    )EOF"};
+    testing::InSequence s;
+    setup(SCRIPT);
+
+    // Setup secure connection if required.
+    EXPECT_CALL(Const(connection_), ssl()).WillOnce(Return(secure ? &ssl_ : nullptr));
+
+    ConnectionWrapper::create(coroutine_->luaState(), &connection_);
+    EXPECT_CALL(*this, testPrint(secure ? "secure" : "plain"));
+    EXPECT_CALL(Const(connection_), ssl()).WillOnce(Return(secure ? &ssl_ : nullptr));
+    EXPECT_CALL(*this, testPrint(secure ? "userdata" : "nil"));
+    start("callMe");
+  }
+
+  NiceMock<Envoy::Network::MockConnection> connection_;
+  NiceMock<Envoy::Ssl::MockConnection> ssl_;
 };
 
 // Basic buffer wrapper methods test.
@@ -222,6 +260,11 @@ TEST_F(LuaMetadataMapWrapperTest, DontFinishIteration) {
   EXPECT_THROW_WITH_MESSAGE(
       start("callMe"), LuaException,
       "[string \"...\"]:5: cannot create a second iterator before completing the first");
+}
+
+TEST_F(LuaConnectionWrapperTest, Secure) {
+  expectSecureConnection(true);
+  expectSecureConnection(false);
 }
 
 } // namespace Lua